thread layout

+++ packages/wasm/as/assembly/w3/invoke.ts
@@ -2,15 +2,15 @@
 
 // Get Invoke Arguments
 @external("w3", "__w3_invoke_args")
-export declare function __w3_invoke_args(name_ptr: i32, args_ptr: i32): void;
+export declare function __w3_invoke_args(name_ptr: usize, args_ptr: usize): void;
 
 // Set Invoke Result
 @external("w3", "__w3_invoke_result")
-export declare function __w3_invoke_result(ptr: i32, len: usize): void;
+export declare function __w3_invoke_result(ptr: usize, len: usize): void;
 
 // Set Invoke Error
 @external("w3", "__w3_invoke_error")
-export declare function __w3_invoke_error(ptr: i32, len: usize): void;
+export declare function __w3_invoke_error(ptr: usize, len: usize): void;
 
 // Keep track of all invokable functions
 type InvokeFunction = (argsBuf: ArrayBuffer) => ArrayBuffer;
@@ -23,11 +23,11 @@ export function w3_add_invoke(name: string, fn: InvokeFunction): void {
 
 // Helper for handling _w3_invoke
 export function w3_invoke(name_size: usize, args_size: usize): bool {
-  const nameBuf = new ArrayBuffer(changetype<i32>(name_size));
-  const argsBuf = new ArrayBuffer(changetype<i32>(args_size));
+  const nameBuf = new ArrayBuffer(changetype<usize>(name_size));
+  const argsBuf = new ArrayBuffer(changetype<usize>(args_size));
   __w3_invoke_args(
-    changetype<i32>(nameBuf),
-    changetype<i32>(argsBuf)
+    changetype<usize>(nameBuf),
+    changetype<usize>(argsBuf)
   );
 
   const name = String.UTF8.decode(nameBuf);
@@ -35,7 +35,7 @@ export function w3_invoke(name_size: usize, args_size: usize): bool {
   if (fn) {
     const result = fn(argsBuf);
     __w3_invoke_result(
-      changetype<i32>(result),
+      changetype<usize>(result),
       result.byteLength
     );
     return true;
@@ -44,7 +44,7 @@ export function w3_invoke(name_size: usize, args_size: usize): bool {
       `Could not find invoke function "${name}"`
     );
     __w3_invoke_error(
-      changetype<i32>(message),
+      changetype<usize>(message),
       message.byteLength
     );
     return false;
